+++ src/components/HeroSelect.js
@@ -22,9 +22,23 @@ class HeroSelect extends Component {
     return (
       <div className="d-flex flex-wrap flex-justify-between">
         <div className="hero-box mb-4">
-          <h5 className="h5 border-bottom pb-2 mb-2">DPS</h5>
+          <h5 className="h5 border-bottom pb-2 mb-2">Flankers</h5>
           <HeroCheckboxList
-            heroes={Hero.byType.DPS}
+            heroes={Hero.byType.Flanker}
+            isAvailable={this.isAvailable}
+            isChecked={this.isChecked}
+            onToggle={onToggle}
+          />
+          <h5 className="h5 border-bottom pb-2 mt-4 mb-2">Main tanks</h5>
+          <HeroCheckboxList
+            heroes={Hero.byType['Main Tank']}
+            isAvailable={this.isAvailable}
+            isChecked={this.isChecked}
+            onToggle={onToggle}
+          />
+          <h5 className="h5 border-bottom pb-2 mb-2 mt-4">Main healers</h5>
+          <HeroCheckboxList
+            heroes={Hero.byType['Main Healer']}
             isAvailable={this.isAvailable}
             isChecked={this.isChecked}
             onToggle={onToggle}
@@ -38,36 +52,30 @@ class HeroSelect extends Component {
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-        </div>
-        <div className="hero-box mb-4">
-          <h5 className="h5 border-bottom pb-2 mb-2">Flankers</h5>
+          <h5 className="h5 border-bottom pb-2 mt-4 mb-2">Off-tanks</h5>
           <HeroCheckboxList
-            heroes={Hero.byType.Flanker}
+            heroes={Hero.byType['Off-tank']}
+            isAvailable={this.isAvailable}
+            isChecked={this.isChecked}
+            onToggle={onToggle}
+          />
+          <h5 className="h5 border-bottom pb-2 mt-4 mb-2">Off-healers</h5>
+          <HeroCheckboxList
+            heroes={Hero.byType['Off-healer']}
             isAvailable={this.isAvailable}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
         </div>
         <div className="hero-box mb-4">
-          <h5 className="h5 border-bottom pb-2 mb-2">Tanks</h5>
+          <h5 className="h5 border-bottom pb-2 mb-2">DPS</h5>
           <HeroCheckboxList
-            heroes={Hero.byType.Tank}
+            heroes={Hero.byType.DPS}
             isAvailable={this.isAvailable}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-        </div>
-        <div className="hero-box mb-4">
-          <h5 className="h5 border-bottom pb-2 mb-2">Healers</h5>
-          <HeroCheckboxList
-            heroes={Hero.byType.Healer}
-            isAvailable={this.isAvailable}
-            isChecked={this.isChecked}
-            onToggle={onToggle}
-          />
-        </div>
-        <div className="hero-box mb-4">
-          <h5 className="h5 border-bottom pb-2 mb-2">Defense</h5>
+          <h5 className="h5 border-bottom pb-2 mt-4 mb-2">Defense</h5>
           <HeroCheckboxList
             heroes={Hero.byType.Defense}
             isAvailable={this.isAvailable}